backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core Web & Tools
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pymongo
requests
pillow
gTTS
groq
python-multipart
fpdf
passlib==1.7.4 
bcrypt==3.2.2
pinecone
pypdf
pyotp
qrcode
cryptography
slowapi
numpy

# ML & AI (CPU-only Optimized)
torch
torchvision
torchaudio
timm
open-clip-torch
